고도화 관련 내용 추가

+++ community_be/src/main/java/hello/hello_spring/service/PostService.java
@@ -49,7 +49,6 @@ public class PostService {
         User user = userRepository.findById(userId).orElseThrow(() -> new RuntimeException("사용자 정보를 찾을 수 없습니다"));
 
         Optional<PostInteraction> interaction = postInteractionRepository.findByPostAndUser(post, user);
-        System.out.println("[debug]: asdfasdfasdf");
         if (interaction.isEmpty()) {
             PostInteraction newInteraction = new PostInteraction();
             newInteraction.setPost(post);
@@ -86,7 +85,7 @@ public class PostService {
                 throw new RuntimeException("해당 게시글을 삭제할 권한이 없습니다.");
             }
 
-            if (post.getPostImageUrl() != null && !post.getPostImageUrl().isEmpty()) {
+            if (post.getPostImageUrl() != null && !post.getPostImageUrl().isEmpty() && !post.getPostImageUrl().equals("/api/images/posts/default.jpg")) {
                 deletePostImage(post.getPostImageUrl());
             }
 
@@ -97,16 +96,13 @@ public class PostService {
             postRepository.save(post);
 
             cleanupDeletedPosts(user);
-            System.out.println("게시글 소프트 삭제 완료");
         }
         catch (DataIntegrityViolationException e) {
             throw new IllegalStateException("데이터 무결성 위반: 게시글 삭제 실패", e);
         } catch (TransactionSystemException e) {
             throw new IllegalStateException("트랜잭션 처리 중 오류 발생", e);
         } catch (Exception e) {
-            System.out.println("aasdfa-------------");
             e.printStackTrace();
-            System.out.println("aasdfa-------------");
             throw new RuntimeException("예상치 못한 오류로 게시글 삭제에 실패했습니다.", e);
         }
     }
@@ -117,13 +113,9 @@ public class PostService {
 
         // ✅ 30일이 지난 Soft Delete된 데이터 조회
         List<Post> postsToDelete = postRepository.findAllByDeletedAtBefore(cutoffDate);
-        System.out.println("[debug in cleanupDeletedPosts] postsToDelete : " + postsToDelete);
         if (!postsToDelete.isEmpty()) {
-
-
             // ✅ 3️⃣ 게시글 삭제
             postRepository.deleteAll(postsToDelete);
-            System.out.println(postsToDelete.size() + "개의 삭제된 게시글을 영구적으로 삭제했습니다.");
         }
     }
 
@@ -136,9 +128,6 @@ public class PostService {
 
         PostInteraction postInteraction = postInteractionRepository.findByPostAndUser(post, user)
                         .orElseThrow(() -> new IllegalStateException("postInteraction이 존재하지 않습니다."));
-        System.out.println("[debug]: asdfasdfasdf");
-        System.out.println("[debug]: " + postInteraction.getPost());
-        System.out.println("[debug] postInteraction.getLine(): " + postInteraction.getLiked());
         if (postInteraction.getPost() == post) {
            if(!postInteraction.getLiked()){
                postInteraction.setLiked(true);
@@ -148,24 +137,21 @@ public class PostService {
                postInteraction.setLiked(false);
                post.setLikeCount(post.getLikeCount() - 1);
            }
-            System.out.println("[debug] post.getLikeCount(): " + post.getLikeCount());
            postInteractionRepository.save(postInteraction);
            this.savePost(post);
         }
     }
 
     private void deletePostImage(String fileUrl) {
+        List<String> defaultImages = List.of("default.jpg", "default_profile.jpg");
         try {
             String baseDirectory = "uploads/posts/"; // 파일이 저장된 기본 폴더
             String fileName = Paths.get(new URI(fileUrl).getPath()).getFileName().toString();
             Path filePath = Paths.get(baseDirectory, fileName);
 
             // 파일 존재 여부 확인 후 삭제
-            if (Files.exists(filePath)) {
+            if (Files.exists(filePath) && !defaultImages.contains(fileName)) {
                 Files.delete(filePath);
-                System.out.println("이미지 삭제 완료: " + filePath);
-            } else {
-                System.out.println("삭제할 파일이 존재하지 않음: " + filePath);
             }
         } catch (Exception e) {
             System.err.println("이미지 삭제 중 오류 발생: " + e.getMessage());
